refactor(engine): hoist filter-pushdown classification into shared/filter_expr (#973)
The HTTP and MCP table providers carried byte-identical classify_filter functions (differing only in a HashSet path) and near-identical supports_filters_pushdown bodies. Hoist the per-expr classifier plus a new classify_filter_pushdown(filters, defined_filters) wrapper into backends/shared/filter_expr.rs, next to its lockstep sibling extract_filter_values (which already builds the same allowed/filter_modes maps). Both providers now delegate via Ok(classify_filter_pushdown(filters, self.table.filters())). The classify_filter unit tests move to filter_expr.rs unchanged. Classification results - and therefore pushdown decisions, query planning, and over-the-wire filter behavior - are identical. Aligns with the backends module guidance to prefer extending the shared module over duplicating.
