+++ crates/coral-engine/src/backends/mcp/provider.rs
@@ -3,12 +3,12 @@ use std::collections::HashMap;
 use std::sync::Arc;
 
 use async_trait::async_trait;
+use coral_spec::ManifestDataType;
 use coral_spec::backends::mcp::McpTableSpec;
-use coral_spec::{FilterMode, ManifestDataType};
 use datafusion::arrow::datatypes::SchemaRef;
 use datafusion::datasource::TableProvider;
 use datafusion::error::{DataFusionError, Result};
-use datafusion::logical_expr::{Expr, Operator, TableProviderFilterPushDown, TableType};
+use datafusion::logical_expr::{Expr, TableProviderFilterPushDown, TableType};
 use datafusion::physical_plan::ExecutionPlan;
 use rmcp::model::JsonObject;
 use serde_json::Value;
@@ -18,7 +18,7 @@ use super::client::McpSourceClient;
 use super::error::McpProviderQueryError;
 use super::fetch::McpFetchPlan;
 use crate::backends::schema_from_columns;
-use crate::backends::shared::filter_expr::{extract_filter_values, literal_to_string};
+use crate::backends::shared::filter_expr::{classify_filter_pushdown, extract_filter_values};
 use crate::backends::shared::json_exec::JsonExec;
 use crate::backends::shared::mapping::convert_items;
 
@@ -76,23 +76,7 @@ impl TableProvider for McpTableProvider {
         &self,
         filters: &[&Expr],
     ) -> Result<Vec<TableProviderFilterPushDown>> {
-        let allowed: std::collections::HashSet<&str> = self
-            .table
-            .filters()
-            .iter()
-            .map(|f| f.name.as_str())
-            .collect();
-        let filter_modes: HashMap<&str, FilterMode> = self
-            .table
-            .filters()
-            .iter()
-            .map(|f| (f.name.as_str(), f.mode))
-            .collect();
-
-        Ok(filters
-            .iter()
-            .map(|expr| classify_filter(expr, &allowed, &filter_modes))
-            .collect())
+        Ok(classify_filter_pushdown(filters, self.table.filters()))
     }
 
     async fn scan(
@@ -281,48 +265,4 @@ fn invalid_filter_value_plan_error(
         "{source_schema}.{table_name} filter '{filter_name}' is declared as {declared} but value \
          '{value}' could not be parsed as {declared}"
     ))
-}
-
-fn classify_filter(
-    expr: &Expr,
-    allowed: &std::collections::HashSet<&str>,
-    filter_modes: &HashMap<&str, FilterMode>,
-) -> TableProviderFilterPushDown {
-    if let Expr::Column(col) = expr
-        && allowed.contains(col.name())
-    {
-        return TableProviderFilterPushDown::Exact;
-    }
-    if let Expr::Not(inner) = expr
-        && let Expr::Column(col) = inner.as_ref()
-        && allowed.contains(col.name())
-    {
-        return TableProviderFilterPushDown::Exact;
-    }
-    if let Expr::IsTrue(inner) | Expr::IsFalse(inner) = expr
-        && let Expr::Column(col) = inner.as_ref()
-        && allowed.contains(col.name())
-    {
-        return TableProviderFilterPushDown::Exact;
-    }
-    if let Expr::BinaryExpr(binary) = expr
-        && binary.op == Operator::Eq
-        && let Expr::Column(col) = binary.left.as_ref()
-        && allowed.contains(col.name())
-        && literal_to_string(binary.right.as_ref()).is_some()
-    {
-        return TableProviderFilterPushDown::Exact;
-    }
-    if let Expr::Like(like) = expr
-        && !like.negated
-        && let Expr::Column(col) = like.expr.as_ref()
-        && allowed.contains(col.name())
-        && literal_to_string(like.pattern.as_ref()).is_some()
-    {
-        let mode = filter_modes.get(col.name()).copied().unwrap_or_default();
-        if matches!(mode, FilterMode::Search | FilterMode::Contains) {
-            return TableProviderFilterPushDown::Inexact;
-        }
-    }
-    TableProviderFilterPushDown::Unsupported
 }
