+++ crates/coral-engine/src/backends/shared/filter_expr.rs
@@ -2,7 +2,7 @@
 
 use std::collections::{HashMap, HashSet};
 
-use datafusion::logical_expr::{Expr, Operator};
+use datafusion::logical_expr::{Expr, Operator, TableProviderFilterPushDown};
 use datafusion::scalar::ScalarValue;
 
 use coral_spec::{FilterMode, FilterSpec};
@@ -24,6 +24,104 @@ pub(crate) fn extract_filter_values(
     }
 
     filters
+}
+
+/// Classifies pushed-down logical expressions for `supports_filters_pushdown`,
+/// mirroring [`extract_filter_values`] arm-for-arm so the pushdown decision and
+/// the value extraction stay in lockstep.
+pub(crate) fn classify_filter_pushdown(
+    filters: &[&Expr],
+    defined_filters: &[FilterSpec],
+) -> Vec<TableProviderFilterPushDown> {
+    let allowed: HashSet<&str> = defined_filters.iter().map(|f| f.name.as_str()).collect();
+    let filter_modes: HashMap<&str, FilterMode> = defined_filters
+        .iter()
+        .map(|f| (f.name.as_str(), f.mode))
+        .collect();
+
+    filters
+        .iter()
+        .map(|expr| classify_filter(expr, &allowed, &filter_modes))
+        .collect()
+}
+
+fn classify_filter(
+    expr: &Expr,
+    allowed: &HashSet<&str>,
+    filter_modes: &HashMap<&str, FilterMode>,
+) -> TableProviderFilterPushDown {
+    if let Expr::BinaryExpr(binary) = expr
+        && binary.op == Operator::And
+    {
+        return classify_filter_conjunction(
+            classify_filter(binary.left.as_ref(), allowed, filter_modes),
+            classify_filter(binary.right.as_ref(), allowed, filter_modes),
+        );
+    }
+    if let Expr::Column(col) = expr
+        && allowed.contains(col.name())
+    {
+        return TableProviderFilterPushDown::Exact;
+    }
+    if let Expr::Not(inner) = expr
+        && let Expr::Column(col) = inner.as_ref()
+        && allowed.contains(col.name())
+    {
+        return TableProviderFilterPushDown::Exact;
+    }
+    if let Expr::IsTrue(inner) | Expr::IsFalse(inner) = expr
+        && let Expr::Column(col) = inner.as_ref()
+        && allowed.contains(col.name())
+    {
+        return TableProviderFilterPushDown::Exact;
+    }
+    if let Expr::BinaryExpr(binary) = expr
+        && binary.op == Operator::Eq
+        && (extract_column_equality(binary.left.as_ref(), binary.right.as_ref(), allowed).is_some()
+            || extract_column_equality(binary.right.as_ref(), binary.left.as_ref(), allowed)
+                .is_some())
+    {
+        return TableProviderFilterPushDown::Exact;
+    }
+    if let Expr::Like(like) = expr
+        && !like.negated
+        && extract_column_like(
+            like.expr.as_ref(),
+            like.pattern.as_ref(),
+            allowed,
+            filter_modes,
+        )
+        .is_some()
+    {
+        // Inexact: the API receives the stripped search/contains term (performance
+        // win) but DataFusion keeps a residual filter to enforce exact
+        // LIKE/ILIKE semantics client-side (correctness win).
+        return TableProviderFilterPushDown::Inexact;
+    }
+    if let Expr::InList(in_list) = expr
+        && !in_list.negated
+        && in_list.list.len() == 1
+        && let Expr::Column(col) = in_list.expr.as_ref()
+        && allowed.contains(col.name())
+        && let Some(literal) = in_list.list.first()
+        && literal_to_string(literal).is_some()
+    {
+        return TableProviderFilterPushDown::Exact;
+    }
+    TableProviderFilterPushDown::Unsupported
+}
+
+fn classify_filter_conjunction(
+    left: TableProviderFilterPushDown,
+    right: TableProviderFilterPushDown,
+) -> TableProviderFilterPushDown {
+    use TableProviderFilterPushDown::{Exact, Inexact, Unsupported};
+
+    match (left, right) {
+        (Unsupported, Unsupported) => Unsupported,
+        (Exact, Exact) => Exact,
+        _ => Inexact,
+    }
 }
 
 fn collect_filter_values(
@@ -315,6 +413,176 @@ mod tests {
         ] {
             let values = extract_filter_values(&[expr], &filters);
             assert!(values.is_empty());
+        }
+    }
+}
+
+#[cfg(test)]
+mod pushdown_classification_tests {
+    use super::classify_filter;
+    use coral_spec::FilterMode;
+    use datafusion::common::Column;
+    use datafusion::logical_expr::{
+        Expr, Operator, TableProviderFilterPushDown, binary_expr, expr::Like, lit,
+    };
+    use std::collections::{HashMap, HashSet};
+    use std::ops::Not;
+
+    fn allowed<'a>(names: &'a [&'a str]) -> HashSet<&'a str> {
+        names.iter().copied().collect()
+    }
+
+    fn modes<'a>(entries: &'a [(&'a str, FilterMode)]) -> HashMap<&'a str, FilterMode> {
+        entries.iter().copied().collect()
+    }
+
+    fn like_expr(col_name: &str, pattern: &str) -> Expr {
+        Expr::Like(Like::new(
+            false,
+            Box::new(col(col_name)),
+            Box::new(lit(pattern)),
+            None,
+            false,
+        ))
+    }
+
+    fn col(name: &str) -> Expr {
+        Expr::Column(Column::from_name(name))
+    }
+
+    #[test]
+    fn like_ignored_for_equality_mode_filter() {
+        let pushdown = classify_filter(
+            &like_expr("status", "%open%"),
+            &allowed(&["status"]),
+            &modes(&[("status", FilterMode::Equality)]),
+        );
+        assert_eq!(pushdown, TableProviderFilterPushDown::Unsupported);
+    }
+
+    #[test]
+    fn strips_wildcards_from_like_pattern() {
+        let pushdown = classify_filter(
+            &like_expr("q", "%deploy runbook%"),
+            &allowed(&["q"]),
+            &modes(&[("q", FilterMode::Contains)]),
+        );
+        assert_eq!(pushdown, TableProviderFilterPushDown::Inexact);
+    }
+
+    #[test]
+    fn contains_filter_also_accepts_equality() {
+        let pushdown = classify_filter(
+            &binary_expr(col("query"), Operator::Eq, lit("deploy")),
+            &allowed(&["query"]),
+            &modes(&[("query", FilterMode::Contains)]),
+        );
+        assert_eq!(pushdown, TableProviderFilterPushDown::Exact);
+    }
+
+    #[test]
+    fn reversed_equality_filter_pushes_down_exactly() {
+        let pushdown = classify_filter(
+            &binary_expr(lit("deploy"), Operator::Eq, col("query")),
+            &allowed(&["query"]),
+            &modes(&[]),
+        );
+        assert_eq!(pushdown, TableProviderFilterPushDown::Exact);
+    }
+
+    #[test]
+    fn single_value_in_list_filter_pushes_down_exactly() {
+        let pushdown = classify_filter(
+            &col("repo").in_list(vec![lit("coral")], false),
+            &allowed(&["repo"]),
+            &modes(&[]),
+        );
+        assert_eq!(pushdown, TableProviderFilterPushDown::Exact);
+    }
+
+    #[test]
+    fn conjunction_of_extractable_filters_pushes_down_exactly() {
+        let pushdown = classify_filter(
+            &binary_expr(col("owner"), Operator::Eq, lit("alice")).and(binary_expr(
+                lit("open"),
+                Operator::Eq,
+                col("status"),
+            )),
+            &allowed(&["owner", "status"]),
+            &modes(&[]),
+        );
+        assert_eq!(pushdown, TableProviderFilterPushDown::Exact);
+    }
+
+    #[test]
+    fn partial_conjunction_pushdown_remains_inexact() {
+        let pushdown = classify_filter(
+            &binary_expr(col("owner"), Operator::Eq, lit("alice")).and(binary_expr(
+                col("unmanaged"),
+                Operator::Eq,
+                lit("open"),
+            )),
+            &allowed(&["owner"]),
+            &modes(&[]),
+        );
+        assert_eq!(pushdown, TableProviderFilterPushDown::Inexact);
+    }
+
+    #[test]
+    fn extracts_like_value_for_contains_mode_filter() {
+        let pushdown = classify_filter(
+            &like_expr("query", "%deploy%"),
+            &allowed(&["query"]),
+            &modes(&[("query", FilterMode::Contains)]),
+        );
+        assert_eq!(pushdown, TableProviderFilterPushDown::Inexact);
+    }
+
+    #[test]
+    fn extracts_like_value_for_legacy_search_mode_filter() {
+        let pushdown = classify_filter(
+            &like_expr("query", "%deploy%"),
+            &allowed(&["query"]),
+            &modes(&[("query", FilterMode::Search)]),
+        );
+        assert_eq!(pushdown, TableProviderFilterPushDown::Inexact);
+    }
+
+    #[test]
+    fn boolean_column_filter_pushes_down_exactly() {
+        let pushdown = classify_filter(&col("descending"), &allowed(&["descending"]), &modes(&[]));
+        assert_eq!(pushdown, TableProviderFilterPushDown::Exact);
+    }
+
+    #[test]
+    fn negated_boolean_column_filter_pushes_down_exactly() {
+        let pushdown = classify_filter(
+            &col("descending").not(),
+            &allowed(&["descending"]),
+            &modes(&[]),
+        );
+        assert_eq!(pushdown, TableProviderFilterPushDown::Exact);
+    }
+
+    #[test]
+    fn boolean_is_true_and_is_false_push_down_exactly() {
+        for expr in [
+            Expr::IsTrue(Box::new(col("descending"))),
+            Expr::IsFalse(Box::new(col("descending"))),
+        ] {
+            let pushdown = classify_filter(&expr, &allowed(&["descending"]), &modes(&[]));
+            assert_eq!(pushdown, TableProviderFilterPushDown::Exact);
+        }
+    }
+
+    #[test]
+    fn null_inclusive_boolean_is_predicates_are_not_pushed_down() {
+        for expr in [
+            Expr::IsNotTrue(Box::new(col("descending"))),
+            Expr::IsNotFalse(Box::new(col("descending"))),
+        ] {
+            let pushdown = classify_filter(&expr, &allowed(&["descending"]), &modes(&[]));
+            assert_eq!(pushdown, TableProviderFilterPushDown::Unsupported);
         }
     }
 }
